requirements.txt
flask==2.3.3
rembg==2.0.50
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
huggingface_hub==0.16.4
